Authentication_roles_05_Web_ITS.ID_1568429-1/src/med_center_project/clinic/authentication.py
from rest_framework_simplejwt.authentication import JWTAuthentication
from rest_framework.exceptions import AuthenticationFailed


class CustomJWTAuthentication(JWTAuthentication):
    def get_user(self, validated_token):
        user = super().get_user(validated_token)
        if not user or not hasattr(user, 'role'):
            raise AuthenticationFailed('Invalid user role')

        return user
    